projects/backend/requirements.txt
Django
djangorestframework
psycopg2
djangorestframework-jwt
